refactor: database:local を単に database モジュールにした
