+++ core/database/.gitignore
