--- core/database/local/schemas/app.kaito_dogi.mybrary.core.database.MybraryDatabase/1.json
@@ -1,52 +1,0 @@
-{
-  "formatVersion": 1,
-  "database": {
-    "version": 1,
-    "identityHash": "a89a781cba60b9d83a276a36446cc13c",
-    "entities": [
-      {
-        "tableName": "draft_memos",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER NOT NULL, `content` TEXT NOT NULL, `from_page` INTEGER, `to_page` INTEGER, PRIMARY KEY(`id`))",
-        "fields": [
-          {
-            "fieldPath": "id",
-            "columnName": "id",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "content",
-            "columnName": "content",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "fromPage",
-            "columnName": "from_page",
-            "affinity": "INTEGER",
-            "notNull": false
-          },
-          {
-            "fieldPath": "toPage",
-            "columnName": "to_page",
-            "affinity": "INTEGER",
-            "notNull": false
-          }
-        ],
-        "primaryKey": {
-          "autoGenerate": false,
-          "columnNames": [
-            "id"
-          ]
-        },
-        "indices": [],
-        "foreignKeys": []
-      }
-    ],
-    "views": [],
-    "setupQueries": [
-      "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, 'a89a781cba60b9d83a276a36446cc13c')"
-    ]
-  }
-}
