+++ core/database/proguard-rules.pro
