+++ core/database/schemas/app.kaito_dogi.mybrary.core.database.MybraryDatabase/1.json
