+++ core/database/src/main/AndroidManifest.xml
