+++ settings.gradle.kts
@@ -19,7 +19,7 @@ include(":app")
 include(":core:api")
 include(":core:common")
 include(":core:data")
-include(":core:database:local")
+include(":core:database")
 include(":core:database:network")
 include(":core:designsystem")
 include(":core:domain")
